# docs/swagger.yaml
basePath: /api/
definitions:
...
  response.Response:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
host: 127.0.0.1:8080
info:
  contact: {}
  description: Swagger docs from sdmedik backend
  title: sdmedik API
  version: "1.0"
paths:
...
  /files/{name}:
    get:
      consumes:
      - application/json
      description: Download a file by its name. Returns file content as attachment.
      parameters:
      - description: File name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/octet-stream
      - image/jpeg
      - image/png
      - application/pdf
      - text/plain
      responses:
        "200":
          description: File content
          headers:
            Content-Disposition:
              description: attachment; filename=example.jpg
              type: string
            Content-Type:
              description: MIME type of the file
              type: string
          schema:
            type: file
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Download file by name
      tags:
      - files
  /files/upload-temporary:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a temporary file to the storage with a specified time-to-live
        (24 hours)
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: File successfully uploaded
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Error reading file or uploading to storage
          schema:
            $ref: '#/definitions/response.Response'
      summary: Upload Temporary File
      tags:
      - files
...
swagger: "2.0"